feat: fast fix for vtrust out of memory
