+++ neurons/validator.py
@@ -35,6 +35,45 @@ from template.base.validator import BaseValidatorNeuron
 load_dotenv()
 
 
+def sync_last_indexed():
+    repo_id = os.getenv("MAIN_REPO_ID")
+    try:
+        dataset = load_dataset(repo_id, split='train', columns=['id'])
+        pipeline = indexing.r.pipeline(transaction=False)
+        batch_size = 10000
+        total_rows = 0
+        indexed_cnt = 0
+        last_indexed = indexing.get(f"{repo_id}:last", None)
+
+        for row in dataset:
+            pipeline.setnx(row['id'], 1)
+            total_rows += 1
+
+            if total_rows % batch_size == 0:
+                try:
+                    indexed_results = pipeline.execute()
+                    indexed_cnt += sum(indexed_results)
+                except Exception as e:
+                    logging.error(f"Failed to index the dataset: {e}")
+                    continue
+                logging.info(f"Processed {total_rows} rows, Indexed {indexed_cnt} new rows.")
+
+        if total_rows % batch_size != 0:
+            indexed_results = pipeline.execute()
+            indexed_cnt += sum(indexed_results)
+
+        last_indexed = dataset[-1]['id']
+        logging.info(f"Total processed rows: {total_rows}, Indexed {indexed_cnt} rows.")
+
+        last_indexed = dataset[-1]['id']
+        flag_key = f"{repo_id}:last"
+        indexing.r.set(flag_key, last_indexed)
+        logging.success(f"Successfully indexed the dataset. Last indexed key: {flag_key}")
+    except Exception as e:
+        logging.error(f"Failed to index the dataset: {e}")
+
+
+
 class Validator(BaseValidatorNeuron):
     """
     Your validator neuron class. You should use this class to define your validator's behavior. In particular, you should replace the forward function with your own logic.
@@ -48,6 +87,7 @@ class Validator(BaseValidatorNeuron):
         super(Validator, self).__init__(config=config)
         self.last_block = self.subtensor.block
         bt.logging.info("load_state()")
+        sync_last_indexed()
         self.load_state()
 
     async def forward(self):
@@ -61,6 +101,7 @@ class Validator(BaseValidatorNeuron):
         """
         # TODO(developer): Rewrite this function based on your protocol definition.
         return await forward(self)
+
 
 # The main function parses the configuration and runs the validator.
 if __name__ == "__main__":
